requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.3
    # via requests
cycler==0.10.0
    # via matplotlib
idna==3.2
    # via requests
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.4.2
    # via -r requirements.in
mypy-extensions==0.4.3
    # via typing-inspect
numpy==1.21.1
    # via
    #   matplotlib
    #   pandas
    #   pandera
packaging==21.0
    # via pandera
pandas==1.3.1
    # via
    #   -r requirements.in
    #   pandera
pandera==0.6.5
    # via -r requirements.in
pillow==8.3.1
    # via matplotlib
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2021.1
    # via pandas
requests==2.26.0
    # via -r requirements.in
six==1.16.0
    # via
    #   cycler
    #   python-dateutil
typing-extensions==3.10.0.0
    # via typing-inspect
typing-inspect==0.7.1
    # via pandera
urllib3==1.26.6
    # via requests
wrapt==1.12.1
    # via pandera
